functions/requirements.txt
functions-framework==3.*
python-docx==0.8.11
pandas==1.5.3
numpy==1.24.3 